Move UserPatchParams out of route module
