+++ app/api/route/schemas/user_schemas.py
@@ -1,11 +1,11 @@
-from datetime import date
-from typing import Any, Optional
+from typing import Any
 
 import marshmallow
 from apiflask import fields
 from marshmallow import fields as marshmallow_fields
 
 from api.db.models.user_models import Role, RoleType, User
+from api.models.user import UserPatchParams
 
 ##############
 # Role Models
@@ -55,41 +55,10 @@ class UserSchema(marshmallow.Schema):
     updated_at = fields.DateTime(dump_only=True)
 
 
-# TODO: dataclass?
-# TODO: maybe extend User?
-class UserPatchParams:
-    # TODO: all of these should be Optional, yeah?
-    first_name: str
-    middle_name: Optional[str]
-    last_name: str
-    phone_number: str
-    date_of_birth: date
-    is_active: bool
-
-    # TODO: fix this
-    roles: Optional[list["Role"]]
-
-    # TODO: this should be hidden somewhere else
-    _fields_set: dict = None
-
-    def __init__(self, data: dict):
-        print("IN init")
-        for key, value in data.items():
-            setattr(self, key, value)
-
-        self._fields_set = data
-
-    # TODO: this should be hidden somewhere
-    # TODO: add comments
-    def get_set_params(self) -> dict:
-        return self._fields_set
-
-
 class UserPostParamsSchema(UserSchema):
     @marshmallow.post_load
     def make_user_post_params(self, data: dict, **kwargs: dict[str, Any]) -> User:
         return User.create_from_data(data)
-
 
 
 # re-use the UserSchema instead of defining those again
